+++ test/test_node.c
@@ -85,27 +85,27 @@ test_string_to_double(void)
 static void
 test_double_to_node(void)
 {
-	const double dbl_test_nums[] = {0.0,
-	                                9.0,
-	                                10.0,
-	                                .01,
-	                                2.05,
-	                                -16.00001,
-	                                5.000000005,
-	                                0.0000000001,
-	                                NAN,
-	                                INFINITY};
+	const double dbl_test_nums[] = { 0.0,
+	                                 9.0,
+	                                 10.0,
+	                                 .01,
+	                                 2.05,
+	                                 -16.00001,
+	                                 5.000000005,
+	                                 0.0000000001,
+	                                 (double)NAN,
+	                                 (double)INFINITY };
 
-	const char* dbl_test_strs[] = {"0.0",
-	                               "9.0",
-	                               "10.0",
-	                               "0.01",
-	                               "2.05",
-	                               "-16.00001",
-	                               "5.00000001",
-	                               "0.0",
-	                               NULL,
-	                               NULL};
+	const char* dbl_test_strs[] = { "0.0",
+	                                "9.0",
+	                                "10.0",
+	                                "0.01",
+	                                "2.05",
+	                                "-16.00001",
+	                                "5.00000001",
+	                                "0.0",
+	                                NULL,
+	                                NULL };
 
 	for (size_t i = 0; i < sizeof(dbl_test_nums) / sizeof(double); ++i) {
 		SerdNode*   node     = serd_new_decimal(dbl_test_nums[i], 8, NULL);
